+++ CHANGELOG.md
@@ -1,6 +1,12 @@
 Yii 2 ActiveRecord Variation extension Change Log
 =================================================
 
+1.0.5 under development
+-----------------------
+
+- Bug #24: Fix ambiguous column error while joining multiple `VariationBehavior::$defaultVariationRelation` (klimov-paul)
+
+
 1.0.4, April 9, 2018
 --------------------
 
@@ -22,7 +28,7 @@ Yii 2 ActiveRecord Variation extension Change Log
 1.0.1, February 10, 2016
 ------------------------
 
-- Bug #13: Preset value for `VariationBehavior::defaultVariationRelation` removed (klimov-paul)
+- Bug #13: Preset value for `VariationBehavior::$defaultVariationRelation` removed (klimov-paul)
 - Bug #12: `VariationBehavior` does not use `ActiveRecord::getRelation()` while retrieving relation instance (klimov-paul)
 - Bug #11: Relation declared via `VariationBehavior::hasDefaultVariationRelation()` does not support `LEFT JOIN` (klimov-paul)